+++ src/Convert/Traverse.hs
@@ -448,13 +448,13 @@ traverseSinglyNestedExprsM exprMapper = em
             pes <- mapM exprMapper $ map snd p
             let p' = zip (map fst p) pes
             return $ Call e' (Args l' p')
-        em (UniOpA     o a e) =
+        em (UniOpA   o a e) =
             exprMapper e >>= return . UniOpA o a
-        em (BinOpA     o a e1 e2) = do
+        em (BinOpA   o a e1 e2) = do
             e1' <- exprMapper e1
             e2' <- exprMapper e2
             return $ BinOpA o a e1' e2'
-        em (MuxA     a e1 e2 e3) = do
+        em (MuxA    a e1 e2 e3) = do
             e1' <- exprMapper e1
             e2' <- exprMapper e2
             e3' <- exprMapper e3