+++ test/basic/expr_attr.sv
@@ -1,4 +1,4 @@
-module top();
-	wire a, b, c, d, e;
-	assign a = b ? (* ternary *) ~^ (* unary *) c : d & (* binary *) e;
+module top;
+    wire a, b, c, d, e;
+    assign a = b ? (* ternary *) ~^ (* unary *) c : d & (* binary *) e;
 endmodule